git-svn-id: https://clsvn.cienet.com.cn/hzkx-caixiang/code/trunk/Phase3/CXServer_3.0/trunk@413 f47374ae-05c8-11e1-8c5d-33d49f259814

+++ bussiness/src/main/java/com/server/cx/service/cx/impl/StatusMGraphicServiceImpl.java
@@ -1,22 +1,25 @@
 package com.server.cx.service.cx.impl;
 
-import java.util.List;
-import org.joda.time.LocalDate;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
 import com.cl.cx.platform.dto.Actions;
 import com.cl.cx.platform.dto.MGraphicDTO;
 import com.google.common.base.Preconditions;
 import com.server.cx.constants.Constants;
 import com.server.cx.dao.cx.StatusTypeDao;
 import com.server.cx.dao.cx.UserStatusMGraphicDao;
+import com.server.cx.entity.cx.GraphicResource;
+import com.server.cx.entity.cx.MGraphic;
 import com.server.cx.entity.cx.StatusType;
 import com.server.cx.entity.cx.UserStatusMGraphic;
 import com.server.cx.model.OperationResult;
 import com.server.cx.service.cx.MGraphicService;
 import com.server.cx.service.cx.StatusTypeService;
 import com.server.cx.service.cx.UserSubscribeGraphicItemService;
+import org.joda.time.LocalDate;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
+import java.util.List;
 
 /**
  * User: yanjianzou
@@ -60,7 +63,7 @@ public class StatusMGraphicServiceImpl extends CheckAndHistoryMGraphicService im
         }
         checkMGraphicIdMustBeNotExists(mGraphicDTO);
 
-        historyPreviousUserCommonMGraphic();
+        historyPreviousMGraphic();
         String mgraphicId = createAndSaveNewUserCommonMGraphic(mGraphicDTO);
         OperationResult operationResult = new OperationResult("createUserStatusMGraphic", Constants.SUCCESS_FLAG);
         if(isImmediate){
@@ -72,15 +75,13 @@ public class StatusMGraphicServiceImpl extends CheckAndHistoryMGraphicService im
 
     @Override
     public OperationResult edit(String imsi, MGraphicDTO mGraphicDTO) {
-        checkAndInitializeContext(imsi, mGraphicDTO);
+        checkAndInitializeUserInfo(imsi);
         checkParameters(imsi, mGraphicDTO);
         checkAndSetUserInfoExists(imsi);
-        
         mGraphicIdMustBeExists(mGraphicDTO);
-        userSubscribeGraphicItemService.checkUserSubscribeGraphicItem(userInfo, mGraphicDTO.getGraphicInfoId());
-        
+
         UserStatusMGraphic mGraphic = userStatusMGraphicDao.findOne(mGraphicDTO.getId());
-        updateMGraphicNameAndSignature(mGraphicDTO, mGraphic);
+        updateMGraphicNameAndSignatureInEditMode(mGraphicDTO, mGraphic);
         userStatusMGraphicDao.save(mGraphic);
 
         return new OperationResult("editUserStatusMGraphic", Constants.SUCCESS_FLAG);
@@ -99,9 +100,16 @@ public class StatusMGraphicServiceImpl extends CheckAndHistoryMGraphicService im
         return new OperationResult("disableUserCommonMGraphic", Constants.SUCCESS_FLAG);
     }
 
-    private void historyPreviousUserCommonMGraphic() {
+    private void historyPreviousMGraphic() {
+        historyPreviousMGraphic(null);
+    }
+
+    private void historyPreviousMGraphic(MGraphic mGraphic){
         List<UserStatusMGraphic> previousUserCommonMGraphics = userStatusMGraphicDao.findByUserInfoAndModeType(userInfo, 5);
         for (UserStatusMGraphic statusMGraphic : previousUserCommonMGraphics) {
+            if(mGraphic !=null && mGraphic.getId().equals(statusMGraphic.getId())){
+                continue;
+            }
             userStatusMGraphicDao.delete(statusMGraphic);
         }
     }
@@ -111,7 +119,10 @@ public class StatusMGraphicServiceImpl extends CheckAndHistoryMGraphicService im
         StatusType statusType = statusTypeDao.findOne(mGraphicDTO.getStatusType());
         deletePreviousMGraphic(statusType);
         UserStatusMGraphic userStatusMGraphic = new UserStatusMGraphic();
-        userStatusMGraphic.setGraphicInfo(graphicInfo);
+        List<GraphicResource> graphicResourceList = graphicInfo.getGraphicResources();
+        if(graphicResourceList !=null && graphicResourceList.size()>0){
+            userStatusMGraphic.setGraphicResource(graphicResourceList.get(0));
+        }
         userStatusMGraphic.setUserInfo(userInfo);
         userStatusMGraphic.setSubscribe(mGraphicDTO.getSubscribe());
         userStatusMGraphic.setStatusType(statusType);
